Clean WOF Set and add tests

+++ Tests/Integration/MSFT_xWindowsFeature.Integration.Tests.ps1
@@ -19,12 +19,6 @@ $script:testEnvironment = Enter-DscResourceTestEnvironment `
     -DscResourceModuleName 'xPSDesiredStateConfiguration' `
     -DscResourceName 'MSFT_xWindowsFeature' `
     -TestType 'Integration'
-
-$script:testFeatureName = 'Telnet-Client'
-$script:testFeatureWithSubFeaturesName = 'ADRMS'
-$script:installStateOfTestFeature
-$script:installStateOfTestWithSubFeatures
-
 <#
     If this is set to $true then the tests that test installing/uninstalling a feature with
     its subfeatures will not run.
@@ -32,21 +26,48 @@ $script:installStateOfTestWithSubFeatures
 $script:skipLongTests = $false
 
 try {
-
-    #Saving the state so we can clean up afterwards
-    $testFeature = Get-WindowsFeature -Name $script:testFeatureName
-    $script:installStateOfTestFeature = $testFeature.Installed
-
-    $testFeatureWithSubFeatures = Get-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
-    $script:installStateOfTestWithSubFeatures = $testFeatureWithSubFeatures.Installed
-
-    $configFile = Join-Path -Path $PSScriptRoot -ChildPath 'MSFT_xWindowsFeature.config.ps1'
-
     Describe 'xWindowsFeature Integration Tests' {
+        BeforeAll {
+            $script:testFeatureName = 'Telnet-Client'
+            $script:testFeatureWithSubFeaturesName = 'RSAT-File-Services'
 
-        $testIncludeAllSubFeature = $false
+            #Saving the state so we can clean up afterwards
+            $testFeature = Get-WindowsFeature -Name $script:testFeatureName
+            $script:installStateOfTestFeature = $testFeature.Installed
 
-        Remove-WindowsFeature -Name $script:testFeatureName
+            $testFeatureWithSubFeatures = Get-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
+            $script:installStateOfTestWithSubFeatures = $testFeatureWithSubFeatures.Installed
+
+            $configFile = Join-Path -Path $PSScriptRoot -ChildPath 'MSFT_xWindowsFeature.config.ps1'
+        }
+
+        AfterAll {
+            # Ensure that features used for testing are re-installed/uninstalled
+            $feature = Get-WindowsFeature -Name $script:testFeatureName
+
+            if ($script:installStateOfTestFeature -and -not $feature.Installed)
+            {
+                Add-WindowsFeature -Name $script:testFeatureName
+            }
+            elseif ( -not $script:installStateOfTestFeature -and $feature.Installed)
+            {
+                Remove-WindowsFeature -Name $script:testFeatureName
+            }
+
+            if (-not $script:skipLongTests)
+            {
+                $feature = Get-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
+
+                if ($script:installStateOfTestWithSubFeatures -and -not $feature.Installed)
+                {
+                    Add-WindowsFeature -Name $script:testFeatureWithSubFeaturesName -IncludeAllSubFeature
+                }
+                elseif ( -not $script:installStateOfTestWithSubFeatures -and $feature.Installed)
+                {
+                    Remove-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
+                }
+            }
+        }
 
         Context "Should Install the Windows Feature: $script:testFeatureName" {
             $configurationName = 'MSFT_xWindowsFeature_InstallFeature'
@@ -63,7 +84,7 @@ try {
                     {
                         . $configFile -ConfigurationName $configurationName
                         & $configurationName -Name $script:testFeatureName `
-                                             -IncludeAllSubFeature $testIncludeAllSubFeature `
+                                             -IncludeAllSubFeature $false `
                                              -Ensure 'Present' `
                                              -OutputPath $configurationPath `
                                              -ErrorAction 'Stop'
@@ -78,7 +99,7 @@ try {
                 It 'Should return the correct configuration' {
                    $currentConfig = Get-DscConfiguration -ErrorAction 'Stop'
                    $currentConfig.Name | Should Be $script:testFeatureName
-                   $currentConfig.IncludeAllSubFeature | Should Be $testIncludeAllSubFeature
+                   $currentConfig.IncludeAllSubFeature | Should Be $false
                    $currentConfig.Ensure | Should Be 'Present'
                 }
 
@@ -112,7 +133,7 @@ try {
                     {
                         . $configFile -ConfigurationName $configurationName
                         & $configurationName -Name $script:testFeatureName `
-                                             -IncludeAllSubFeature $testIncludeAllSubFeature `
+                                             -IncludeAllSubFeature $false `
                                              -Ensure 'Absent' `
                                              -OutputPath $configurationPath `
                                              -ErrorAction 'Stop'
@@ -127,7 +148,7 @@ try {
                 It 'Should return the correct configuration' {
                    $currentConfig = Get-DscConfiguration -ErrorAction 'Stop'
                    $currentConfig.Name | Should Be $script:testFeatureName
-                   $currentConfig.IncludeAllSubFeature | Should Be $testIncludeAllSubFeature
+                   $currentConfig.IncludeAllSubFeature | Should Be $false
                    $currentConfig.Ensure | Should Be 'Absent'
                 }
 
@@ -155,58 +176,43 @@ try {
 
             $logPath = Join-Path -Path $TestDrive -ChildPath 'InstallSubFeatureTest.log'
 
-            try
+            if (-not $script:skipLongTests)
             {
-                if (-not $script:skipLongTests)
-                {
-                    # Ensure that the feature is not already installed
-                    Remove-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
-                }
-
-                It 'Should compile without throwing' -Skip:$script:skipLongTests {
-                    {
-                        . $configFile -ConfigurationName $configurationName
-                        & $configurationName -Name $script:testFeatureWithSubFeaturesName `
-                                             -IncludeAllSubFeature $true `
-                                             -Ensure 'Present' `
-                                             -OutputPath $configurationPath `
-                                             -ErrorAction 'Stop'
-                        Start-DscConfiguration -Path $configurationPath -ErrorAction 'Stop' -Wait -Force
-                    } | Should Not Throw
-                }
-
-                It 'Should be able to call Get-DscConfiguration without throwing' -Skip:$script:skipLongTests {
-                    { Get-DscConfiguration -ErrorAction 'Stop' } | Should Not Throw
-                }
-                
-                It 'Should return the correct configuration' -Skip:$script:skipLongTests {
-                   $currentConfig = Get-DscConfiguration -ErrorAction 'Stop'
-                   $currentConfig.Name | Should Be $script:testFeatureWithSubFeaturesName
-                   $currentConfig.IncludeAllSubFeature | Should Be $true
-                   $currentConfig.Ensure | Should Be 'Present'
-                }
-
-                It 'Should be Installed (includes check for subFeatures)' -Skip:$script:skipLongTests {
-                    $feature = Get-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
-                    $feature.Installed | Should Be $true
-
-                    foreach ($subFeatureName in $feature.SubFeatures)
-                    {
-                        $subFeature = Get-WindowsFeature -Name $subFeatureName
-                        $subFeature.Installed | Should Be $true
-                    }
-                }
-
+                # Ensure that the feature is not already installed
+                Remove-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
             }
-            finally
-            {
-                if (Test-Path -Path $logPath) {
-                    Remove-Item -Path $logPath -Recurse -Force
-                }
 
-                if (Test-Path -Path $configurationPath)
+            It 'Should compile without throwing' -Skip:$script:skipLongTests {
                 {
-                    Remove-Item -Path $configurationPath -Recurse -Force
+                    . $configFile -ConfigurationName $configurationName
+                    & $configurationName -Name $script:testFeatureWithSubFeaturesName `
+                                            -IncludeAllSubFeature $true `
+                                            -Ensure 'Present' `
+                                            -OutputPath $configurationPath `
+                                            -ErrorAction 'Stop'
+                    Start-DscConfiguration -Path $configurationPath -ErrorAction 'Stop' -Wait -Force
+                } | Should Not Throw
+            }
+
+            It 'Should be able to call Get-DscConfiguration without throwing' -Skip:$script:skipLongTests {
+                { Get-DscConfiguration -ErrorAction 'Stop' } | Should Not Throw
+            }
+                
+            It 'Should return the correct configuration' -Skip:$script:skipLongTests {
+                $currentConfig = Get-DscConfiguration -ErrorAction 'Stop'
+                $currentConfig.Name | Should Be $script:testFeatureWithSubFeaturesName
+                $currentConfig.IncludeAllSubFeature | Should Be $true
+                $currentConfig.Ensure | Should Be 'Present'
+            }
+
+            It 'Should be Installed (includes check for subFeatures)' -Skip:$script:skipLongTests {
+                $feature = Get-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
+                $feature.Installed | Should Be $true
+
+                foreach ($subFeatureName in $feature.SubFeatures)
+                {
+                    $subFeature = Get-WindowsFeature -Name $subFeatureName
+                    $subFeature.Installed | Should Be $true
                 }
             }
         }
@@ -217,52 +223,37 @@ try {
 
             $logPath = Join-Path -Path $TestDrive -ChildPath 'UninstallSubFeatureTest.log'
 
-            try
-            {
-                It 'Should compile without throwing' -Skip:$script:skipLongTests {
-                    {
-                        . $configFile -ConfigurationName $configurationName
-                        & $configurationName -Name $script:testFeatureWithSubFeaturesName `
-                                             -IncludeAllSubFeature $true `
-                                             -Ensure 'Absent' `
-                                             -OutputPath $configurationPath `
-                                             -ErrorAction 'Stop'
-                        Start-DscConfiguration -Path $configurationPath -ErrorAction 'Stop' -Wait -Force
-                    } | Should Not Throw
-                }
-
-                It 'Should be able to call Get-DscConfiguration without throwing' -Skip:$script:skipLongTests {
-                    { Get-DscConfiguration -ErrorAction 'Stop' } | Should Not Throw
-                }
-                
-                It 'Should return the correct configuration' -Skip:$script:skipLongTests  {
-                   $currentConfig = Get-DscConfiguration -ErrorAction 'Stop'
-                   $currentConfig.Name | Should Be $script:testFeatureWithSubFeaturesName
-                   $currentConfig.IncludeAllSubFeature | Should Be $false
-                   $currentConfig.Ensure | Should Be 'Absent'
-                }
-
-                It 'Should not be installed (includes check for subFeatures)' -Skip:$script:skipLongTests {
-                    $feature = Get-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
-                    $feature.Installed | Should Be $false
-
-                    foreach ($subFeatureName in $feature.SubFeatures)
-                    {
-                        $subFeature = Get-WindowsFeature -Name $subFeatureName
-                        $subFeature.Installed | Should Be $false
-                    }
-                }
-
-            }
-            finally
-            {
-                if (Test-Path -Path $logPath) {
-                    Remove-Item -Path $logPath -Recurse -Force
-                }
-
-                if (Test-Path -Path $configurationPath)
+            It 'Should compile without throwing' -Skip:$script:skipLongTests {
                 {
-                    Remove-Item -Path $configurationPath -Recurse -Force
+                    . $configFile -ConfigurationName $configurationName
+                    & $configurationName -Name $script:testFeatureWithSubFeaturesName `
+                                            -IncludeAllSubFeature $true `
+                                            -Ensure 'Absent' `
+                                            -OutputPath $configurationPath `
+                                            -ErrorAction 'Stop'
+                    Start-DscConfiguration -Path $configurationPath -ErrorAction 'Stop' -Wait -Force
+                } | Should Not Throw
+            }
+
+            It 'Should be able to call Get-DscConfiguration without throwing' -Skip:$script:skipLongTests {
+                { Get-DscConfiguration -ErrorAction 'Stop' } | Should Not Throw
+            }
+                
+            It 'Should return the correct configuration' -Skip:$script:skipLongTests  {
+                $currentConfig = Get-DscConfiguration -ErrorAction 'Stop'
+                $currentConfig.Name | Should Be $script:testFeatureWithSubFeaturesName
+                $currentConfig.IncludeAllSubFeature | Should Be $false
+                $currentConfig.Ensure | Should Be 'Absent'
+            }
+
+            It 'Should not be installed (includes check for subFeatures)' -Skip:$script:skipLongTests {
+                $feature = Get-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
+                $feature.Installed | Should Be $false
+
+                foreach ($subFeatureName in $feature.SubFeatures)
+                {
+                    $subFeature = Get-WindowsFeature -Name $subFeatureName
+                    $subFeature.Installed | Should Be $false
                 }
             }
         }
@@ -270,28 +261,5 @@ try {
 }
 finally
 {
-    # Ensure that features used for testing are re-installed/uninstalled
-    if ($script:installStateOfTestFeature)
-    {
-        Add-WindowsFeature -Name $script:testFeatureName
-    }
-    else
-    {
-        Remove-WindowsFeature -Name $script:testFeatureName
-    }
-
-    if (-not $script:skipLongTests)
-    {
-        if ($script:installStateOfTestWithSubFeatures)
-        {
-            Add-WindowsFeature -Name $script:testFeatureWithSubFeaturesName -IncludeAllSubFeature
-        }
-        else
-        {
-            Remove-WindowsFeature -Name $script:testFeatureWithSubFeaturesName
-        }
-    }
     Exit-DscResourceTestEnvironment -TestEnvironment $script:testEnvironment
 }
-
-
